migration/requirements.txt
# Requirements for PostgreSQL to MongoDB Migration

# Database drivers
psycopg2-binary==2.9.7
pymongo==4.5.0

# Additional utilities
python-dotenv==1.0.0
click==8.1.7

# For data validation and processing
jsonschema==4.19.1